+++ CHANGELOG.md
@@ -10,10 +10,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial creation of the changelog file.
 
+## [2025-03-19]
+
+####
+- Updated version to 0.0.16 and added new feat to get a GitLab Group variable and create/update a Group Variable
+
 ## [2025-03-18]
 
 #### Fixed
-- Updated version to 0.0.16 and and enhance project name existence check logic.
+- Updated version to 0.0.16 and enhance project name existence check logic.
 
 ## [2025-03-07]